refactor(compile): extract native-addon detection to collect_modules/native_addon.rs
Keeps collect_modules.rs under the 2000-line file-size cap (the wall fixes
pushed it from 1988 to 2031). Moves the self-contained compile-package
Node-native-addon probe group (package-root resolution + binding.gyp /
prebuilds / gypfile / loader-dep / *.node markers + the refusal) into a
sibling module; behavior unchanged.
